Added Singleton

+++ Assets/Scripts/Player/Player.cs
@@ -6,9 +6,9 @@ namespace Player
 {
     public class Player : MonoBehaviour
     {
-        private int _health;
+        private float _health;
 
-        public void TakeDamage(int damage)
+        public void TakeDamage(float damage)
         {
             _health -= damage;
         }
